server/requirements.txt
flask
flask-sqlalchemy
requests
gunicorn
python-dotenv
user-agents
